+++ Sources/KeyboardKit/Extensions/UIInputViewController+Orientation.swift
@@ -13,9 +13,6 @@ extension UIInputViewController {
     /**
      Get the current screen orientation. If no window can be
      resolved `portrait` is returned.
-     
-     Note that you probably shouldn't use screen orientation,
-     but rather use the controller traits instead.
      */
     var screenOrientation: UIInterfaceOrientation {
         view.window?.screen.orientation ?? .portrait